// display-tables/src/interfaces/amountsWithoutLosses.ts
export interface AmountsWithoutLossesInterface {
  paymentAmountTotalDay: number;
  paymentAmountTotalNight: number;
  paymentAmountMinerDay: number;
  paymentAmountMinerNight: number;
  paymentAmountHousingWorkerDay: number;
  paymentAmountHousingWorkerNight: number;
  paymentAmountOrelDay: number;
  paymentAmountOrelNight: number;
}