// display-tables/src/interfaces/amountsOfLosses.ts
export interface AmountsOfLossesInterface {
  lossPaymentAmountTotalDay: number;
  lossPaymentAmountTotalNight: number;
  lossPaymentAmountMinerDay: number;
  lossPaymentAmountMinerNight: number;
  lossPaymentAmountHousingWorkerDay: number;
  lossPaymentAmountHousingWorkerNight: number;
  lossPaymentAmountOrelDay: number;
  lossPaymentAmountOrelNight: number;
}